Specify dtypes

+++ quantile_forest/tests/test_utils.py
@@ -41,7 +41,7 @@ def test_generate_unsampled_indices():
 
 def test_group_indices_by_value():
     # Check grouping indices by value.
-    inputs = np.array([1, 3, 2, 2, 5, 4, 5, 5])
+    inputs = np.array([1, 3, 2, 2, 5, 4, 5, 5], dtype=np.int64)
 
     actual_indices, actual_values = group_indices_by_value(inputs)
     expected_indices = np.array([1, 2, 3, 4, 5])
@@ -60,9 +60,9 @@ def test_group_indices_by_value():
 def test_map_indices_to_leaves():
     # Check mapping of indices to leaf nodes.
     y_train_leaves = np.zeros((3, 1, 3), dtype=np.int64)
-    bootstrap_indices = np.array([[1], [2], [3], [4], [5]])
-    leaf_indices = np.array([1, 2])
-    leaf_values_list = [np.array([0, 1, 2]), np.array([3, 4])]
+    bootstrap_indices = np.array([[1], [2], [3], [4], [5]], dtype=np.int64)
+    leaf_indices = np.array([1, 2], dtype=np.int64)
+    leaf_values_list = [np.array([0, 1, 2], dtype=np.int64), np.array([3, 4], dtype=np.int64)]
 
     actual = map_indices_to_leaves(
         y_train_leaves,
